--- front/src/pages/Home.tsx
@@ -1,11 +1,0 @@
-import Feed from "../components/Feed";
-import ReviewModalButton from "../components/ReviewModalButton";
-
-export default function Home() {
-  return (
-    <>
-      <ReviewModalButton />
-      <Feed />
-    </>
-  );
-}
